refactor: settings code structure to new format
As the team, we decided to refactor our folder structure to denote
private modules with underscore prefix. This change just changes the
folder structure without other functional changes.
