+++ tensorboard/webapp/settings/BUILD
@@ -1,65 +1,21 @@
-load("//tensorboard/defs:defs.bzl", "tf_ng_module", "tf_ts_library")
+load("//tensorboard/defs:defs.bzl", "tf_ng_module", "tf_ng_web_test_suite")
 
 package(default_visibility = ["//tensorboard:internal"])
 
 tf_ng_module(
     name = "settings",
     srcs = [
-        "dialog_component.ts",
-        "polymer_interop_container.ts",
-        "settings_button_component.ts",
         "settings_module.ts",
     ],
-    assets = [
-        "dialog_component.css",
-    ],
     deps = [
-        "//tensorboard/webapp:tb_polymer_interop_types",
-        "//tensorboard/webapp/angular:expect_angular_material_button",
-        "//tensorboard/webapp/angular:expect_angular_material_checkbox",
-        "//tensorboard/webapp/angular:expect_angular_material_dialog",
-        "//tensorboard/webapp/angular:expect_angular_material_icon",
-        "//tensorboard/webapp/angular:expect_angular_material_input",
-        "//tensorboard/webapp/core",
-        "//tensorboard/webapp/core/actions",
-        "//tensorboard/webapp/core/store",
-        "@npm//@angular/common",
+        "//tensorboard/webapp/settings/_views",
         "@npm//@angular/core",
-        "@npm//@angular/forms",
-        "@npm//@ngrx/store",
-        "@npm//rxjs",
     ],
 )
 
-tf_ts_library(
-    name = "test_lib",
-    testonly = True,
-    srcs = [
-        "polymer_interop_test.ts",
-        "settings_test.ts",
-    ],
+tf_ng_web_test_suite(
+    name = "karma_test",
     deps = [
-        ":settings",
-        "//tensorboard/webapp/angular:expect_angular_cdk_overlay",
-        "//tensorboard/webapp/angular:expect_angular_core_testing",
-        "//tensorboard/webapp/angular:expect_angular_material_button",
-        "//tensorboard/webapp/angular:expect_angular_material_checkbox",
-        "//tensorboard/webapp/angular:expect_angular_material_dialog",
-        "//tensorboard/webapp/angular:expect_angular_material_icon",
-        "//tensorboard/webapp/angular:expect_angular_material_input",
-        "//tensorboard/webapp/angular:expect_angular_platform_browser_animations",
-        "//tensorboard/webapp/angular:expect_angular_platform_browser_dynamic_testing",
-        "//tensorboard/webapp/angular:expect_ngrx_store_testing",
-        "//tensorboard/webapp/core/actions",
-        "//tensorboard/webapp/core/store",
-        "//tensorboard/webapp/core/testing",
-        "//tensorboard/webapp/testing:mat_icon",
-        "@npm//@angular/common",
-        "@npm//@angular/compiler",
-        "@npm//@angular/core",
-        "@npm//@angular/forms",
-        "@npm//@angular/platform-browser",
-        "@npm//@ngrx/store",
-        "@npm//@types/jasmine",
+        "//tensorboard/webapp/settings/_views:test_lib",
     ],
 )
