+++ tensorboard/webapp/settings/_views/settings_module.ts
@@ -1,0 +1,47 @@
+/* Copyright 2019 The TensorFlow Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
+import {MatButtonModule} from '@angular/material/button';
+import {MatCheckboxModule} from '@angular/material/checkbox';
+import {MatDialogModule} from '@angular/material/dialog';
+import {MatIconModule} from '@angular/material/icon';
+import {MatInputModule} from '@angular/material/input';
+
+import {SettingsButtonComponent} from './settings_button_component';
+import {SettingsDialogComponent} from './dialog_component';
+import {SettingsPolymerInteropContainer} from './polymer_interop_container';
+
+@NgModule({
+  declarations: [
+    SettingsButtonComponent,
+    SettingsDialogComponent,
+    SettingsPolymerInteropContainer,
+  ],
+  exports: [SettingsButtonComponent, SettingsPolymerInteropContainer],
+  entryComponents: [SettingsDialogComponent],
+  imports: [
+    CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
+    MatButtonModule,
+    MatCheckboxModule,
+    MatDialogModule,
+    MatIconModule,
+    MatInputModule,
+  ],
+})
+export class SettingsModule {}
